# much faster toStraightAlpha + other stuff