not only is the conversion much faster, it is better due to being able to use round(). this does imply all output images change a bit.

before:
```
fill_rgba .......................... 0.382 ms      0.387 ms    ±0.004  x1000
isOpaque ........................... 0.399 ms      0.404 ms    ±0.009  x1000
toStraightAlpha ................... 11.180 ms     11.446 ms    ±0.173   x435
```

after:
```
fill_rgba .......................... 0.289 ms      0.310 ms    ±0.015  x1000
isOpaque ........................... 0.164 ms      0.170 ms    ±0.009  x1000
toStraightAlpha .................... 4.418 ms      4.523 ms    ±0.127  x1000
```